finished lesson 3

+++ controllers/api/comment-routes.js
@@ -11,25 +11,29 @@ router.get('/', (req, res) => {
 });
 
 router.post('/', (req, res) => {
-  // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
-  Comment.create({
-    comment_text: req.body.comment_text,
-    user_id: req.session.user_id,
-    post_id: req.body.post_id
-  })
+  // check the seesion
+  if (req.session) {
+    // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
+    Comment.create({
+      comment_text: req.body.comment_text,
+      user_id: req.session.user_id,
+      post_id: req.body.post_id
+    })
     .then(dbCommentData => res.json(dbCommentData))
     .catch(err => {
       console.log(err);
       res.status(400).json(err);
     });
+  }
 });
 
 router.delete('/:id', (req, res) => {
-  Comment.destroy({
-    where: {
-      id: req.params.id
-    }
-  })
+  if (req.session) {
+    Comment.destroy({
+      where: {
+        id: req.params.id
+      }
+    })
     .then(dbCommentData => {
       if (!dbCommentData) {
         res.status(404).json({ message: 'No comment found with this id!' });
@@ -41,6 +45,7 @@ router.delete('/:id', (req, res) => {
       console.log(err);
       res.status(500).json(err);
     });
+  }
 });
 
 module.exports = router;
